+++ src/thread.cpp
@@ -193,6 +193,26 @@ void thread::tick() {
             m_stack.back() /= top;
             break;
         }
+        case UDV: {
+            if (m_flags.warnings) {
+                if (m_stack.size() < 2) UNLIKELY {
+                    cerr << "Warning: Attempt to divide from stack with fewer than 2 elements.\n";
+                }
+                if (!m_stack.empty() && m_stack.back() == INT24_C(0)) UNLIKELY {
+                    cerr << "Warning: Attempted division by zero.\n";
+                }
+            }
+
+            int24_t top = m_stack.back();
+            m_stack.pop_back();
+            int24_t second = m_stack.back();
+
+            uint32_t unsigned_first = static_cast<uint32_t>(top) & 0x00ff'ffffU;
+            uint32_t unsigned_second = static_cast<uint32_t>(second) & 0x00ff'ffffU;
+
+            m_stack.back() = static_cast<int24_t>(unsigned_second / unsigned_first);
+            break;
+        }
         case MOD: {
             if (m_flags.warnings) {
                 if (m_stack.size() < 2) UNLIKELY {
@@ -364,6 +384,16 @@ void thread::tick() {
                 exit(EXIT_SUCCESS);
             }
             break;
+        case PTU:
+            EMPTY_PROTECT("print from") {
+                cout << (static_cast<uint32_t>(m_stack.back()) & 0x00ff'ffffU) << '\n';
+            }
+
+            if (m_flags.pipekill && ferror(stdout)) {
+                cerr << flush;
+                exit(EXIT_SUCCESS);
+            }
+            break;
         case SKP:
             advance();
             m_status = status::idle;
@@ -374,15 +404,15 @@ void thread::tick() {
                 break;
             }
 
-            int24_t top = m_stack.back();
+            size_t index = static_cast<size_t>(m_stack.back()) & 0x00ff'ffffU;
             m_stack.pop_back();
 
-            if (m_flags.warnings && (top < INT24_C(0) || m_stack.size() < static_cast<size_t>(top) + 1)) UNLIKELY {
+            if (m_flags.warnings && m_stack.size() < index + 1) UNLIKELY {
                 cerr << "Warning: Attempt to index out of stack bounds (size = " << m_stack.size()
-                     << ", index = " << top << ")\n";
+                     << ", index = " << index << ")\n";
             }
 
-            size_t i = m_stack.size() - static_cast<size_t>(top) - 1;
+            size_t i = m_stack.size() - index - 1;
             m_stack.push_back(m_stack[i]);
 
             break;