+++ tests/cli/cat/index.sh
@@ -13,3 +13,6 @@ folder=$(dirname "$0")
 output=$($TRILANGLE "${folder}/cat.trg" <<<"$text")
 # Okay so on Windows the newline characters are different. Isn't this fun?
 test "$text" = "${output//$'\r'/}"
+
+x=$((0x912345))
+test $x = "$($TRILANGLE "${folder}/uint.trg" <<<$x)"